// server/enums/ItemIds.ts
export enum ItemIds {
  SWORD = 0, // ну сделай в строки index хм ItemIds["SWORD"] will work ItemIds[id.toUpperCase()]
  PICK = 1,
  FUR = 2,
  PICK_GOLD = 3,
  PICK_DIAMOND = 4,
  SWORD_GOLD = 5,
  SWORD_DIAMOND = 6,
  HAND = 7,
  PICK_WOOD = 8,
  PIRATE_SWORD = 9,
  EARMUFFS = 10,
  COAT = 11,
  WOOD_SPEAR = 12,
  SPEAR = 13,
  GOLD_SPEAR = 14,
  DIAMOND_SPEAR = 15,
  DRAGON_SPEAR = 16,
  LAVA_SPEAR = 17,
  CRAB_SPEAR = 18,
  REIDITE_SWORD = 19,
  DIAMOND_PROTECTION = 20,
  AMETHYST_PROTECTION = 21,
  REIDITE_PROTECTION = 22,
  EXPLORER_HAT = 23,
  PIRATE_HAT = 24,
  STONE_HELMET = 25,
  GOLD_HELMET = 26,
  DIAMOND_HELMET = 27,
  BOOK = 28,
  BAG = 29,
  SWORD_AMETHYST = 30,
  PICK_AMETHYST = 31,
  PICK_REIDITE = 32,
  AMETHYST_SPEAR = 33,
  REIDITE_SPEAR = 34,
  HAMMER = 35,
  HAMMER_GOLD = 36,
  HAMMER_DIAMOND = 37,
  HAMMER_AMETHYST = 38,
  HAMMER_REIDITE = 39,
  CAP_SCARF = 40,
  CHRISTMAS_HAT = 41,
  ELF_HAT = 42,
  AMETHYST_HELMET = 43,
  REIDITE_HELMET = 44,
  SUPER_HAMMER = 45,
  SHOVEL = 46,
  SUPER_DIVING_SUIT = 47,
  DIVING_MASK = 48,
  WATERING_CAN_FULL = 49,
  SHOVEL_GOLD = 50,
  SHOVEL_DIAMOND = 51,
  SHOVEL_AMETHYST = 52,
  PITCHFORK = 53,
  PITCHFORK2 = 54,
  SPANNER = 55,
  MACHETE = 56,
  SWORD_WOOD = 57,
  WOOD_HELMET = 58,
  DRAGON_HELMET = 59,
  LAVA_HELMET = 60,
  CROWN_CRAB = 61,
  DRAGON_SWORD = 62,
  LAVA_SWORD = 63,
  WOOD_BOW = 64,
  STONE_BOW = 65,
  GOLD_BOW = 66,
  DIAMOND_BOW = 67,
  AMETHYST_BOW = 68,
  REIDITE_BOW = 69,
  DRAGON_BOW = 70,
  WOOD_SHIELD = 71,
  STONE_SHIELD = 72,
  GOLD_SHIELD = 73,
  DIAMOND_SHIELD = 74,
  AMETHYST_SHIELD = 75,
  REIDITE_SHIELD = 76,
  CROWN_GREEN = 77,
  CROWN_ORANGE = 78,
  CROWN_BLUE = 79,
  TURBAN1 = 80,
  TURBAN2 = 81,
  PILOT_HELMET = 82,
  HOOD = 83,
  PEASANT = 84,
  WINTER_HOOD = 85,
  WINTER_PEASANT = 86,
  FLOWER_HAT = 87,
  FUR_HAT = 88,
  SADDLE = 89,
  WITCH = 90,
  NIMBUS = 91,
  WAND1 = 92,
  WAND2 = 93,
  FIREFLY = 94,
  WOOD_ARROW = 95,
  STONE_ARROW = 96,
  GOLD_ARROW = 97,
  DIAMOND_ARROW = 98,
  AMETHYST_ARROW = 99,
  REIDITE_ARROW = 100,
  DRAGON_ARROW = 101,
  STONE = 102,
  WOOD = 103,
  PLANT = 104,
  GOLD = 105,
  DIAMOND = 106,
  FIRE = 107,
  WORKBENCH = 108,
  SEED = 109,
  MEAT = 110,
  COOKED_MEAT = 111,
  BIG_FIRE = 112,
  FURNACE = 113,
  PAPER = 114,
  AMETHYST = 115,
  AMETHYST_WALL = 116,
  AMETHYST_SPIKE = 117,
  AMETHYST_DOOR = 118,
  BRIDGE = 119,
  SAND = 120,
  BOTTLE_FULL = 121,
  BOTTLE_EMPTY = 122,
  KRAKEN_SKIN = 123,
  WATERING_CAN = 124,
  FLOUR = 125,
  WHEAT_SEED = 126,
  COOKIE = 127,
  WILD_WHEAT = 128,
  WINDMILL = 129,
  CAKE = 130,
  FOODFISH = 131,
  FOODFISH_COOKED = 132,
  SCALES = 133,
  GROUND = 134,
  PLOT = 135,
  ICE = 136,
  BREAD = 137,
  BREAD_OVEN = 138,
  SANDWICH = 139,
  FUR_WINTER = 140,
  BLUE_CORD = 141,
  LOCK = 142,
  DRAGON_HEART = 143,
  LAVA_HEART = 144,
  RESURRECTION = 145,
  EMERALD_MACHINE = 146,
  EXTRACTOR_MACHINE_STONE = 147,
  EXTRACTOR_MACHINE_GOLD = 148,
  EXTRACTOR_MACHINE_DIAMOND = 149,
  EXTRACTOR_MACHINE_AMETHYST = 150,
  EXTRACTOR_MACHINE_REIDITE = 151,
  LOCKPICK = 152,
  TOTEM = 153,
  SPIKE = 154,
  CORD = 155,
  WALL = 156,
  STONE_WALL = 157,
  GOLD_WALL = 158,
  DIAMOND_WALL = 159,
  WOOD_DOOR = 160,
  CHEST = 161,
  STONE_SPIKE = 162,
  GOLD_SPIKE = 163,
  DIAMOND_SPIKE = 164,
  STONE_DOOR = 165,
  GOLD_DOOR = 166,
  DIAMOND_DOOR = 167,
  FUR_WOLF = 168,
  GEMME_GREEN = 169,
  GEMME_ORANGE = 170,
  GEMME_BLUE = 171,
  SPECIAL_FUR = 172,
  SPECIAL_FUR_2 = 173,
  BUCKET_FULL = 174,
  BUCKET_EMPTY = 175,
  WELL = 176,
  SIGN = 177,
  DRAGON_CUBE = 178,
  DRAGON_ORB = 179,
  LAVA_CUBE = 180,
  LAVA_ORB = 181,
  PUMPKIN_SEED = 182,
  PUMPKIN = 183,
  ROOF = 184,
  GARLIC_SEED = 185,
  GARLIC = 186,
  THORNBUSH_SEED = 187,
  THORNBUSH = 188,
  BANDAGE = 189,
  CRAB_STICK = 190,
  CRAB_LOOT = 191,
  BED = 192,
  SUGAR_CAN = 193,
  CANDY = 194,
  GARLAND = 195,
  REIDITE = 196,
  FLAME = 197,
  CARROT_SEED = 198,
  CARROT = 199,
  TOMATO_SEED = 200,
  TOMATO = 201,
  WATERMELON_SEED = 202,
  WATERMELON = 203,
  ALOE_VERA_SEED = 204,
  ALOE_VERA = 205,
  WOOD_DOOR_SPIKE = 206,
  STONE_DOOR_SPIKE = 207,
  GOLD_DOOR_SPIKE = 208,
  DIAMOND_DOOR_SPIKE = 209,
  AMETHYST_DOOR_SPIKE = 210,
  REIDITE_WALL = 211,
  REIDITE_DOOR = 212,
  REIDITE_SPIKE = 213,
  REIDITE_DOOR_SPIKE = 214,
  WOOD_TOWER = 215,
  PENGUIN_FEATHER = 216,
  BOAT = 217,
  SLED = 218,
  MOUNT_BOAR = 219,
  CRAB_BOSS = 220,
  BABY_DRAGON = 221,
  BABY_LAVA = 222,
  HAWK = 223,
  PLANE = 224,
  HAWK_FEATHER = 225,
  VULTURE_FEATHER = 226,
  CACTUS = 227,
  EMERALD = 228,
  PITCHFORK_PART = 229,
  PILOT_GLASSES = 230,
  FUR_BOAR = 231,
  SANDWORM_JUICE = 232,
  BABY_MAMMOTH = 233,
  FUR_MAMMOTH = 234,
}